6.5 NLP-LLM-GENAI/4.0 LANGCHAIN/requirements.txt
langchain
langchain-core

langchain-openai
openai

langchain-google-genai
google-generativeai

langchain-huggingface
transformers
huggingface-hub

python-dotenv

numpy
scikit-learn